requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
Bottleneck==1.3.4
cachetools==5.0.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==37.0.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.1
fonttools==4.25.0
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
greenlet==1.1.2
h11==0.13.0
httplib2==0.20.4
idna==3.3
importlib-resources==5.2.0
ipykernel==6.9.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.4.2
lxml==4.8.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.8
numexpr==2.8.1
numpy==1.21.5
oauthlib==3.2.0
outcome==1.2.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.2
plotly==5.9.0
prometheus-client==0.13.1
prompt-toolkit==3.0.20
protobuf==3.20.1
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.4
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pywin32==302
pywinpty==2.0.2
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
selenium==4.3.0
Send2Trash==1.8.0
setuptools==61.2.0
sip==6.5.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
stack-data==0.2.0
tenacity==8.0.1
terminado==0.13.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.1
widgetsnbextension==3.5.2
wincertstore==0.2
wordcloud==1.8.1
wsproto==1.1.0
zipp==3.7.0
